refactor(ui): create common taskSearch component
